Create basic entity component system
There is probably some issues regarding heap allocation and pointer chasing still present in the code, which is why this will be committed to the experimental branch for now.  The GameObject class works, as it is able to create a player-like entity, but I think there is definitely room for improvement.